chore(internals): rename getVersion to getVersionFromPackageJson
More explicit name to not be confused by getVersion which was now renamed to getEnginesVersion
