+++ packages/internals/src/engine-commands/errorHelpers.ts
@@ -1,5 +1,5 @@
 import { formatTable } from '../utils/formatTable'
-import { version } from '../utils/getVersion'
+import { version } from '../utils/getVersionFromPackageJson'
 
 /**
  * Adds `Prisma CLI Version : x.x.x` at the bottom of the error output.
