+++ packages/internals/src/index.ts
@@ -55,7 +55,7 @@ export { formatTable } from './utils/formatTable'
 export { getCommandWithExecutor } from './utils/getCommandWithExecutor'
 export type { EnvPaths } from './utils/getEnvPaths'
 export { getEnvPaths } from './utils/getEnvPaths'
-export { version } from './utils/getVersion'
+export { version } from './utils/getVersionFromPackageJson'
 export { handlePanic } from './utils/handlePanic'
 export { isCi } from './utils/isCi'
 export { isCurrentBinInstalledGlobally } from './utils/isCurrentBinInstalledGlobally'
