+++ packages/internals/src/utils/getVersionFromPackageJson.ts
